copy-paste example from geeks3d.com
http://www.geeks3d.com/20120102/programming-tutorial-simple-x11-x-window-code-sample-for-linux-and-mac-os-x/

this code does not specifically state an author or copyright holder
nor does it state a license.

However, this is specifically posted as an example with a tutorial
as to how to build it, so I think any reasonable person can assume
that the intent is for people to copy and use it.